handleScript
